refactor:sqlit recovery setting
