--- src/ui/gui_widgets/gui_menmonic_test.h
@@ -1,29 +1,0 @@
-#ifdef COMPILE_SIMULATOR
-#define _UNUSED
-#else
-#define _UNUSED __attribute__((unused))
-#endif
-
-static const char _UNUSED *test20list1 = "upgrade romp academic acne bracelet verdict reject satisfy main extra carpet pacific quantity sniff society earth havoc froth orbit";
-static const char _UNUSED *test20list2 = "upgrade romp academic agree diminish radar findings random bulb jewelry dismiss syndrome starting webcam insect ranked usher inside rocky";
-static const char _UNUSED *test20list3 = "upgrade romp academic amazing alien huge step news elder execute making living grant talent midst earth marvel nail isolate";
-static const char _UNUSED *testlist1 = "object recover academic acne ajar wolf home usher chest favorite enjoy pile herd desert desert express obesity leaves ticket equation costume punish damage kind stadium earth flash paper worthy gums snake species large";
-static const char _UNUSED *testlist2 = "object recover academic agree aviation walnut upstairs burning slavery volume coding dramatic rescue custody scandal switch mouse benefit smell exhaust wrist lecture activity paper forward ladybug ticket snapshot owner science quick priority lily";
-static const char _UNUSED *testlist3 = "object recover academic amazing angel license lily beyond worthy sympathy judicial loan location gray equip cricket wrist mountain holy bracelet kitchen drink ivory premium formal pile remove knit screw group lawsuit pregnant beaver";
-// static const char *testlist4 = "object recover academic arcade academic luck budget warn being lamp brother capacity fact kitchen remind recover vampire alien enjoy beam plot artist flame idle standard ting animal arena permit shrimp calcium soul";
-// static const char *testlist5 = "object recover academic axle alarm body reject crowd aluminum parcel union jerky argue moisture taught move bumpy cover quantity biology deny aluminum payment jewelry agency climate tension umbrella market weapon artwork kitchen";
-// static const char *testlist6 = "object recover academic bishop arcade believe cluster round tracks adorn maximum texture submit main axis alive black replace news benefit threaten disease marvel pencil maiden gray facility oral unusual favorite hour admit";
-// static const char *testlist7 = "object recover academic breathe amuse hanger famous shame staff costume snake float total salt ordinary usher laundry drift downtown erode grin lips video plastic lungs retailer aquatic decision rebuild visual patrol advance";
-// static const char *testlist8 = "object recover academic cage afraid hesitate skunk class dish mama prospect satisfy decent sunlight holy gesture hush process become exchange prospect race suitable industry ajar woman rich eclipse spend fraction rumor knife";
-// static const char *testlist9 = "object recover academic ceiling advance climate reward exchange payment typical iris adorn curious twice easel garden corner cage club romantic quantity evidence primary toxic drove preach vintage trash faint pumps vanish increase";
-// static const char *testlist10 = "object recover academic client anatomy category civil photo exceed expand born news width vanish smirk vegan company member aquatic reject gray jury length document public tricycle engage nuclear deny calcium midst artwork";
-// static const char *testlist11 = "object recover academic counter apart forward flash prisoner ladle greatest equation chubby season repeat elder airline exact amuse problem tenant toxic secret wine diminish pumps deploy bumpy desktop harvest rich friendly apart";
-// static const char *testlist12 = "object recover academic crystal alcohol game sheriff emperor minister seafood cover reject axle payroll pupal national elephant obesity liquid tofu decrease voting spew triumph drift imply platform emphasis amuse benefit duke imply";
-// static const char *testlist13 = "object recover academic debut acne slow hazard mule lungs bucket tactics surprise exact evil twice repair plastic raisin husband terminal plastic various charity terminal unwrap public ecology pencil center category database swimming";
-// static const char *testlist14 = "object recover academic desire ancestor shrimp vintage knife grumpy repeat rhythm enjoy observe findings burning course photo cluster force ticket lawsuit salt aspect earth gesture valid vanish spine elevator material evoke patrol";
-// static const char *testlist15 = "object recover academic disease award puny legal injury flame merchant teammate wrap physics camera mule tadpole slap pancake warn response woman language island density goat counter prize ivory bundle deadline leader peaceful";
-// static const char *testlist16 = "object recover academic dress airline rainbow careful oven regular depend leaf grill identify agree grin exclude salt depend source require critical exercise flea unknown unhappy hairy buyer amount isolate length thank swing";
-static const char _UNUSED *g_bpi39Test = "unfair enlist digital broccoli track peasant echo mother little exist street tragic";
-static const char _UNUSED *g_bpi3918Test = "popular night utility detect anchor defy box federal bundle exact matter pumpkin company target battle clean metal maze";
-// const static char *g_bpi3924Test = "legal genuine summer person couch copy square clog gloom office gadget find citizen ensure pipe pluck fiscal defense across steak weird pig chicken unaware";
-static const char _UNUSED *g_bpi3924Test = "risk dwarf cactus forum meat road casino near federal husband north owner journey spider adapt fantasy congress box excess crazy travel essay upgrade upset";
