+++ src/ui/gui_widgets/setting/gui_seed_check_widgets.c
@@ -11,605 +11,236 @@
 #include "keystore.h"
 #include "gui_setting_widgets.h"
 #include "gui_lock_widgets.h"
-#include "gui_menmonic_test.h"
 #include "presetting.h"
 #include "assert.h"
 #include "gui_qr_hintbox.h"
 #include "motor_manager.h"
 #include "fingerprint_process.h"
 #include "gui_qrcode_widgets.h"
+#include "gui_mnemonic_input.h"
 #ifndef COMPILE_SIMULATOR
 #include "sha256.h"
 #include "keystore.h"
 #else
 #include "simulator_model.h"
-#define FP_SUCCESS_CODE                                                             (0)
-#define RECOGNIZE_UNLOCK                                                            (0)
-#define RECOGNIZE_OPEN_SIGN                                                         (1)
-#define RECOGNIZE_SIGN                                                              (2)
-#define NO_ENCRYPTION                                                               (0)
-#define AES_KEY_ENCRYPTION                                                          (1)
-#define RESET_AES_KEY_ENCRYPTION                                                    (2)
-#define FINGERPRINT_EN_SING_ERR_TIMES                                               (5)
 #endif
 
-static void RecognizeSuccussHandler(lv_timer_t *timer);
-static void RecognizeFailHandler(lv_timer_t *timer);
-static void FingerButtonHandler(lv_event_t *e);
-static void FingerUnlockDeviceHandler(lv_event_t *e);
-static void FingerDeleteDialogsHandler(lv_event_t *e);
-static void FingerDeleteHandler(lv_event_t *e);
-static void FingerCancelDeleteHandler(lv_event_t *e);
+void RebootHandler(lv_event_t *e);
+static void ImportPhraseWordsHandler(lv_event_t *e);
+static void ImportShareNextSliceHandler(lv_event_t *e);
 
-static lv_obj_t *g_fpAddCont = NULL;
-static lv_obj_t *g_imgFinger = NULL;
-static lv_obj_t *g_arcProgress = NULL;
-static lv_obj_t *g_fpRegLabel = NULL;
-static lv_obj_t *g_imgSignFinger = NULL;
-static lv_obj_t *g_verifyFingerCont = NULL;
-static lv_obj_t *g_fpSingerSwitch = NULL;
-static lv_obj_t *g_fingerManagerContainer = NULL;
-static lv_obj_t *g_labelSignFailed = NULL;
-static lv_obj_t *g_fpUnlockSwitch = NULL;
-static lv_timer_t *g_fpRecognizeTimer = NULL;
-static bool g_firstAddFingerFlag = false;
-static uint8_t g_fpEnSignCnt = 0;
-static uint8_t g_deleteFingerIndex = 0;
+static lv_obj_t *g_buttonCont = NULL;           // next buton cont
+static uint8_t g_inputWordsCnt = 0;
+static MnemonicKeyBoard_t *g_recoveryMkb;       // recovery mnemonic keyboard
+static KeyBoard_t *g_recoveryPhraseKb;          // recovery keyboard
+static lv_obj_t *g_recoveryTitle; // recovery title label
 
-uint8_t GuiGetFingerSettingIndex(void)
-{
-    if (GetRegisteredFingerNum() == 0) {
-        g_firstAddFingerFlag = true;
-        return DEVICE_SETTING_FINGER_ADD_ENTER;
-    } else {
-        return DEVICE_SETTING_FINGER_MANAGER;
-    }
-}
-
-void CancelVerifyFingerHandler(lv_event_t *e)
+void ResetSeedCheckImportHandler(lv_event_t *e)
 {
     lv_event_code_t code = lv_event_get_code(e);
     if (code == LV_EVENT_CLICKED) {
-        FpCancelCurOperate();
-        GUI_DEL_OBJ(g_verifyFingerCont)
+        ClearMnemonicKeyboard(g_recoveryMkb, &g_recoveryMkb->currentId);
+        GuiClearKeyBoard(g_recoveryPhraseKb);
     }
 }
 
-void CancelCurFingerHandler(lv_event_t *e)
+void GuiWalletSeedCheckClearKb(void)
 {
-    lv_event_code_t code = lv_event_get_code(e);
-    if (code == LV_EVENT_CLICKED) {
-        FpCancelCurOperate();
-        GuiEmitSignal(SIG_SETUP_VIEW_TILE_PREV, NULL, 0);
-    }
+    ClearMnemonicKeyboard(g_recoveryMkb, &g_recoveryMkb->currentId);
+    GuiClearMnemonicKeyBoard(g_recoveryMkb);
+    GuiDelMnemonicKeyBoard(g_recoveryMkb);
 }
 
-void GuiFingerManagerDestruct(void *obj, void *param)
+void GuiWalletSeedCheckClearObject(void)
 {
-    if (g_fpRecognizeTimer != NULL) {
-        lv_timer_del(g_fpRecognizeTimer);
-        g_fpRecognizeTimer = NULL;
-    }
-    ClearSecretCache();
+    CLEAR_OBJECT(g_recoveryMkb);
 }
 
-void GuiFpVerifyDestruct(void)
+void GuiWalletRecoveryDestruct(void *obj, void *param)
 {
-    GUI_DEL_OBJ(g_verifyFingerCont)
+    // todo
+    lv_obj_add_flag(g_recoveryPhraseKb->cont, LV_OBJ_FLAG_HIDDEN);
+    lv_obj_add_flag((lv_obj_t *)obj, LV_OBJ_FLAG_HIDDEN);
+    lv_obj_add_flag(g_buttonCont, LV_OBJ_FLAG_HIDDEN);
+    // lv_obj_del((lv_obj_t *)obj);
+    g_recoveryPhraseKb = NULL;
+    lv_obj_del(g_recoveryMkb->nextButton);
+    g_buttonCont = NULL;
+    // lv_obj_del(obj);
 }
 
-void FingerSignHandler(lv_event_t *e)
-{
-    lv_event_code_t code = lv_event_get_code(e);
-
-    if (code == LV_EVENT_CLICKED) {
-        bool en = lv_obj_has_state(g_fpSingerSwitch, LV_STATE_CHECKED);
-        if (en) {
-            lv_obj_clear_state(g_fpSingerSwitch, LV_STATE_CHECKED);
-            UpdateFingerSignFlag(GetCurrentAccountIndex(), !en);
-            SetFingerManagerInfoToSE();
-            lv_event_send(g_fpSingerSwitch, LV_EVENT_VALUE_CHANGED, NULL);
-        } else {
-            g_verifyFingerCont = GuiCreateHintBox(lv_scr_act(), 480, 428, true);
-            lv_obj_t *cont = g_verifyFingerCont;
-            lv_obj_t *label = GuiCreateIllustrateLabel(cont, "Verify Fingerprint");
-            lv_obj_set_style_text_opa(label, LV_OPA_56, LV_PART_MAIN);
-            lv_obj_align(label, LV_ALIGN_DEFAULT, 36, 402);
-
-            lv_obj_t *img = GuiCreateImg(cont, &imgClose);
-            GuiButton_t table[] = {
-                {
-                    .obj = img,
-                    .align = LV_ALIGN_DEFAULT,
-                    .position = {14, 14},
-                }
-            };
-            lv_obj_t *button = GuiCreateButton(cont, 64, 64, table, 1, CancelVerifyFingerHandler, NULL);
-            lv_obj_align(button, LV_ALIGN_DEFAULT, 384, 394);
-
-            g_imgSignFinger = GuiCreateImg(cont, &imgYellowFinger);
-            lv_obj_align(g_imgSignFinger, LV_ALIGN_BOTTOM_MID, 0, -178);
-
-            g_labelSignFailed = GuiCreateLabel(cont, "Verify Failed. Please try Again!");
-            lv_obj_set_style_text_color(g_labelSignFailed, RED_COLOR, LV_PART_MAIN);
-            lv_obj_add_flag(g_labelSignFailed, LV_OBJ_FLAG_HIDDEN);
-            lv_obj_align(g_labelSignFailed, LV_ALIGN_BOTTOM_MID, 0, -100);
-
-            lv_obj_t *arc = GuiCreateArc(cont);
-            lv_obj_align(arc, LV_ALIGN_BOTTOM_MID, 0, -154);
-            FpRecognize(RECOGNIZE_OPEN_SIGN);
-        }
-    }
-}
-
-void GuiSettingDealFingerRecognize(void *param)
-{
-    if (g_verifyFingerCont == NULL) {
-        return;
-    }
-    uint8_t errCode = *(uint8_t *)param;
-    if (errCode == FP_SUCCESS_CODE) {
-        g_fpEnSignCnt = 0;
-        GuiClearQrcodeSignCnt();
-        lv_img_set_src(g_imgSignFinger, &imgGreenFinger);
-        lv_obj_add_flag(g_labelSignFailed, LV_OBJ_FLAG_HIDDEN);
-        UpdateFingerSignFlag(GetCurrentAccountIndex(), true);
-        g_fpRecognizeTimer = lv_timer_create(RecognizeSuccussHandler, 500, NULL);
-    } else {
-        g_fpEnSignCnt++;
-        lv_obj_clear_flag(g_labelSignFailed, LV_OBJ_FLAG_HIDDEN);
-        lv_img_set_src(g_imgSignFinger, &imgRedFinger);
-        if (g_fpEnSignCnt >= FINGERPRINT_EN_SING_ERR_TIMES) {
-            GUI_DEL_OBJ(g_verifyFingerCont)
-            ClearSecretCache();
-            GuiEmitSignal(SIG_SETUP_VIEW_TILE_PREV, NULL, 0);
-        } else {
-            g_fpRecognizeTimer = lv_timer_create(RecognizeFailHandler, 1000, NULL);
-            FpRecognize(RECOGNIZE_OPEN_SIGN);
-        }
-    }
-}
-
-static void ClearFingerErrorStateView(lv_event_t *e)
-{
-    lv_event_code_t code = lv_event_get_code(e);
-
-    if (code == LV_EVENT_CLICKED) {
-        GuiEmitSignal(SIG_SETUP_VIEW_TILE_PREV, NULL, 0);
-        lv_arc_set_value(g_arcProgress, 0);
-        lv_img_set_src(g_imgFinger, &imgWhiteFinger);
-        lv_obj_set_style_arc_color(g_arcProgress, WHITE_COLOR, LV_PART_INDICATOR);
-        lv_label_set_text(g_fpRegLabel, "");
-        for (int i = 0; i < 3; i++) {
-            if (GetFingerRegisteredStatus(i) == 0) {
-                RegisterFp(i);
-                break;
-            }
-        }
-    }
-}
-
-static void FirstAddFingerToManagerView(lv_event_t *e)
-{
-    lv_event_code_t code = lv_event_get_code(e);
-
-    if (code == LV_EVENT_CLICKED) {
-        GuiSettingCloseToTargetTileView(2);
-        uint8_t walletIndex = DEVICE_SETTING_FINGER_MANAGER;
-        GuiEmitSignal(SIG_SETUP_VIEW_TILE_NEXT, &walletIndex, sizeof(walletIndex));
-    }
-}
-
-static void AddFingerToManagerView(lv_event_t *e)
-{
-    lv_event_code_t code = lv_event_get_code(e);
-
-    if (code == LV_EVENT_CLICKED) {
-        GuiSettingCloseToTargetTileView(3);
-    }
-}
-
-void GuiWalletFingerAddFpWidget(lv_obj_t *parent, bool success)
+void GuiWalletRecoveryMethodCheck(lv_obj_t *parent)
 {
     lv_obj_set_style_bg_opa(parent, LV_OPA_0, LV_PART_SCROLLBAR | LV_STATE_SCROLLED);
     lv_obj_set_style_bg_opa(parent, LV_OPA_0, LV_PART_SCROLLBAR | LV_STATE_DEFAULT);
+    lv_obj_t *button;
 
-    lv_obj_t *img = NULL;
-    lv_obj_t *label = NULL;
-    lv_obj_t *btn = NULL;
-    lv_obj_t *desc = NULL;
+    if (PassphraseExist(GetCurrentAccountIndex()) == true) {
+        lv_obj_t *cont = GuiCreateContainerWithParent(parent, 408, 514);
+        lv_obj_set_style_bg_color(cont, WHITE_COLOR, LV_PART_MAIN);
+        lv_obj_set_style_bg_opa(cont, LV_OPA_10 + LV_OPA_2, LV_PART_MAIN);
+        lv_obj_set_style_radius(cont, 24, LV_PART_MAIN);
+        lv_obj_align(cont, LV_ALIGN_TOP_MID, 0, 156 - GUI_MAIN_AREA_OFFSET);
 
-    if (success) {
-        img = GuiCreateImg(parent, &imgSuccess);
-        label = GuiCreateLittleTitleLabel(parent, "Add Successfully");
-        btn = GuiCreateBtn(parent, "Done");
-    } else {
-        img = GuiCreateImg(parent, &imgFailed);
-        label = GuiCreateLittleTitleLabel(parent, "Failed to Add");
-        btn = GuiCreateBtn(parent, "Try Again");
-        desc = GuiCreateNoticeLabel(parent, "Failed to add fingerprint, please use another finger and try again");
-        lv_obj_align(desc, LV_ALIGN_BOTTOM_MID, 0, -404);
-        lv_obj_set_style_text_align(desc, LV_TEXT_ALIGN_CENTER, LV_PART_MAIN);
-    }
-    lv_obj_align(img, LV_ALIGN_BOTTOM_MID, 0, -548);
-    lv_obj_align(label, LV_ALIGN_BOTTOM_MID, 0, -476);
-    lv_obj_set_size(btn, 408, 66);
-    lv_obj_align(btn, LV_ALIGN_BOTTOM_MID, 0, -24);
-    lv_obj_set_style_bg_color(btn, ORANGE_COLOR, LV_PART_MAIN);
-    if (success) {
-        if (g_firstAddFingerFlag == true) {
-            g_firstAddFingerFlag = false;
-            lv_obj_add_event_cb(btn, FirstAddFingerToManagerView, LV_EVENT_CLICKED, NULL);
-        } else {
-            lv_obj_add_event_cb(btn, AddFingerToManagerView, LV_EVENT_CLICKED, NULL);
-        }
-    } else {
-        lv_obj_add_event_cb(btn, ClearFingerErrorStateView, LV_EVENT_CLICKED, NULL);
-    }
-}
+        lv_obj_t *imgLock = GuiCreateImg(cont, &imgWalletLock);
+        lv_obj_align(imgLock, LV_ALIGN_TOP_MID, 0, 36);
 
-void GuiFingerAddDestruct(void *obj, void *param)
-{
-    g_fpAddCont = NULL;
-    GUI_DEL_OBJ(g_imgFinger)
-    GUI_DEL_OBJ(g_arcProgress)
-    GUI_DEL_OBJ(g_fpRegLabel)
-}
+        lv_obj_t *disableTitle = GuiCreateTextLabel(cont, _("Disable Passphrase Wallet"));
+        lv_obj_align(disableTitle, LV_ALIGN_TOP_MID, 0, 140);
+        lv_obj_set_style_text_align(disableTitle, LV_TEXT_ALIGN_CENTER, LV_PART_MAIN);
 
-void GuiSettingFingerRegisterSuccess(void *param)
-{
-    if (g_fpAddCont == NULL) {
+        lv_obj_t *disableDesc = GuiCreateNoticeLabel(cont, "You are presently using a passphrase-protected wallet. Prior to proceeding with the seed phrase verification process, please restart your device without entering the passphrase.");
+        lv_obj_align(disableDesc, LV_ALIGN_TOP_MID, 0, 188);
+        lv_obj_set_style_text_align(disableDesc, LV_TEXT_ALIGN_CENTER, LV_PART_MAIN);
+        lv_label_set_long_mode(disableDesc, LV_LABEL_LONG_WRAP);
+        lv_obj_set_width(disableDesc, 336);
+
+        lv_obj_t *restartLabel = GuiCreateTextLabel(cont, _("Restart Now"));
+        lv_obj_set_style_text_color(restartLabel, ORANGE_COLOR, LV_PART_MAIN);
+        GuiButton_t restartTable[] = {
+            {
+                .obj = restartLabel,
+                .align = LV_ALIGN_CENTER,
+                .position = {0, 0},
+            },
+        };
+        button = GuiCreateButton(cont, 198, 66, restartTable, NUMBER_OF_ARRAYS(restartTable), RebootHandler, NULL);
+        lv_obj_align(button, LV_ALIGN_BOTTOM_MID, 0, -26);
         return;
     }
-    uint8_t step = *(uint8_t *)param;
-    lv_arc_set_value(g_arcProgress, step);
-    lv_img_set_src(g_imgFinger, &imgYellowFinger);
-    lv_obj_set_style_arc_color(g_arcProgress, ORANGE_COLOR, LV_PART_INDICATOR);
-    lv_label_set_text(g_fpRegLabel, "");
-    if (step == 8) {
-        uint8_t walletIndex = DEVICE_SETTING_FINGER_ADD_SUCCESS;
-        GuiEmitSignal(SIG_SETUP_VIEW_TILE_NEXT, &walletIndex, sizeof(walletIndex));
-    }
-}
-
-void GuiSettingFingerRegisterFail(void *param)
-{
-    if (g_fpAddCont == NULL) {
-        return;
-    }
-    if (param == NULL) {
-        uint8_t walletIndex = DEVICE_SETTING_FINGER_ADD_OUT_LIMIT;
-        GuiEmitSignal(SIG_SETUP_VIEW_TILE_NEXT, &walletIndex, sizeof(walletIndex));
-    } else {
-        uint8_t errCode = *(uint8_t *)param;
-        char *text = NULL;
-        lv_img_set_src(g_imgFinger, &imgRedFinger);
-        printf("errCode = %#x\n", errCode);
-        lv_obj_set_style_arc_color(g_arcProgress, RED_COLOR, LV_PART_INDICATOR);
-        if (errCode == 0x93) { // finger registered
-            text = "Duplicate finger, Please change the other finger.";
-        } else {
-            text = (char *)GetFpErrorMessage(errCode);
-        }
-        lv_label_set_text(g_fpRegLabel, text);
-    }
-}
-
-void GuiWalletFingerAddWidget(lv_obj_t *parent)
-{
-    lv_obj_set_style_bg_opa(parent, LV_OPA_0, LV_PART_SCROLLBAR | LV_STATE_SCROLLED);
-    lv_obj_set_style_bg_opa(parent, LV_OPA_0, LV_PART_SCROLLBAR | LV_STATE_DEFAULT);
-    g_fpAddCont = parent;
-    lv_obj_t *label = NULL;
-    label = GuiCreateTitleLabel(parent, "Add Fingerprint");
-    lv_obj_align(label, LV_ALIGN_TOP_LEFT, 36, 12);
-
-    label = GuiCreateNoticeLabel(parent, "Place your finger on the sensor and remove it when you feel a vibration");
-    lv_obj_align(label, LV_ALIGN_TOP_LEFT, 36, 72);
-
-    g_imgFinger = GuiCreateImg(parent, &imgWhiteFinger);
-    lv_obj_align(g_imgFinger, LV_ALIGN_BOTTOM_MID, 0, -264);
-
-    g_arcProgress = GuiCreateArc(parent);
-    lv_arc_set_range(g_arcProgress, 0, 8);
-    lv_obj_align(g_arcProgress, LV_ALIGN_BOTTOM_MID, 0, -240);
-
-    g_fpRegLabel = GuiCreateLabel(parent, "");
-    lv_obj_set_style_text_color(g_fpRegLabel, RED_COLOR, LV_PART_MAIN);
-    lv_obj_set_style_text_align(g_fpRegLabel, LV_TEXT_ALIGN_CENTER, 0);
-    lv_obj_align(g_fpRegLabel, LV_ALIGN_BOTTOM_MID, 0, -156);
-
-    for (int i = 0; i < 3; i++) {
-        if (GetFingerRegisteredStatus(i) == 0) {
-            RegisterFp(i);
-            break;
-        }
-    }
-#ifdef COMPILE_SIMULATOR
-    label = GuiCreateTextLabel(parent, "Register complete");
-    GuiButton_t table[1] = {
-        {
-            .obj = label,
-            .align = LV_ALIGN_BOTTOM_MID,
-            .position = {0, -15},
-        }
-    };
-    static uint32_t walletSetting = DEVICE_SETTING_FINGER_ADD_SUCCESS;
-    lv_obj_t *button = NULL;
-    button = GuiCreateButton(parent, 408, 66, table, 1, WalletSettingHandler, &walletSetting);
-    lv_obj_align(button, LV_ALIGN_BOTTOM_MID, 0, -24);
-#endif
-}
-
-void GuiWalletFingerManagerWidget(lv_obj_t *parent)
-{
-    lv_obj_set_style_bg_opa(parent, LV_OPA_0, LV_PART_SCROLLBAR | LV_STATE_SCROLLED);
-    lv_obj_set_style_bg_opa(parent, LV_OPA_0, LV_PART_SCROLLBAR | LV_STATE_DEFAULT);
-
-    lv_obj_t *container = GuiCreateContainerWithParent(parent, 480, 656);
-    lv_obj_set_align(container, LV_ALIGN_BOTTOM_MID);
-    g_fingerManagerContainer = container;
-    GuiFingerMangerStructureCb(NULL, NULL);
-}
-
-void GuiFingerMangerStructureCb(void *obj, void *param)
-{
-    lv_obj_t *label = NULL;
-    lv_obj_t *imgArrow = NULL;
-    lv_obj_t *button = NULL;
-    lv_obj_t *container = g_fingerManagerContainer;
-    uint32_t curPositionY = 0;
-    g_fpEnSignCnt = 0;
-
-    static uint32_t g_curDealFingerIndex[3] = {0, 1, 2};
-    static uint32_t walletSetting[3] = {
-        DEVICE_SETTING_FINGER_ADD_ENTER,
-        DEVICE_SETTING_FINGER_DELETE,
-        DEVICE_SETTING_FINGER_SET_PATTERN
+    static uint8_t walletSetting[2] = {
+        DEVICE_SETTING_RECOVERY_SINGLE_PHRASE,
+        DEVICE_SETTING_RECOVERY_SHARE_PHRASE
     };
 
-    if (lv_obj_get_child_cnt(container) > 0) {
-        lv_obj_clean(container);
-    }
-    label = GuiCreateTextLabel(container, "Unlock Device");
-    g_fpUnlockSwitch = GuiCreateSwitch(container);
+    lv_obj_t *label = GuiCreateTextLabel(parent, _("seed_check_single_phrase"));
+    lv_obj_t *imgArrow = GuiCreateImg(parent, &imgArrowRight);
 
-    if (GetFingerUnlockFlag() == 1) {
-        lv_obj_add_state(g_fpUnlockSwitch, LV_STATE_CHECKED);
-    } else {
-        lv_obj_clear_state(g_fpUnlockSwitch, LV_STATE_CHECKED);
-    }
-    lv_obj_clear_flag(g_fpUnlockSwitch, LV_OBJ_FLAG_CLICKABLE);
-
-    GuiButton_t table[] = {
+    GuiButton_t table[2] = {
         {
             .obj = label,
             .align = LV_ALIGN_LEFT_MID,
             .position = {24, 0},
         },
         {
-            .obj = g_fpUnlockSwitch,
-            .align = LV_ALIGN_LEFT_MID,
-            .position = {376, 0},
-        },
-    };
-    button = GuiCreateButton(container, 456, 84, table, NUMBER_OF_ARRAYS(table), FingerUnlockDeviceHandler, NULL);
-    if (GetRegisteredFingerNum() == 0) {
-        lv_obj_clear_flag(button, LV_OBJ_FLAG_CLICKABLE);
-        lv_obj_set_style_bg_opa(button, LV_OPA_10, LV_PART_MAIN);
-    }
-    lv_obj_align(button, LV_ALIGN_DEFAULT, 12, curPositionY);
-    curPositionY += 96;
-
-    label = GuiCreateTextLabel(container, "Sign Transactions");
-    g_fpSingerSwitch = GuiCreateSwitch(container);
-    if (GetFingerSignFlag() == 1) {
-        lv_obj_add_state(g_fpSingerSwitch, LV_STATE_CHECKED);
-    } else {
-        lv_obj_clear_state(g_fpSingerSwitch, LV_STATE_CHECKED);
-    }
-    lv_obj_clear_flag(g_fpSingerSwitch, LV_OBJ_FLAG_CLICKABLE);
-    table[0].obj = label;
-    table[1].obj = g_fpSingerSwitch;
-    button = GuiCreateButton(container, 456, 84, table, NUMBER_OF_ARRAYS(table), FingerSignHandler, NULL);
-    if (GetRegisteredFingerNum() == 0) {
-        lv_obj_clear_flag(button, LV_OBJ_FLAG_CLICKABLE);
-        lv_obj_set_style_bg_opa(button, LV_OPA_10, LV_PART_MAIN);
-    }
-    lv_obj_align(button, LV_ALIGN_DEFAULT, 12, curPositionY);
-    curPositionY += 96;
-
-    // bool en = lv_obj_has_state(g_fpSingerSwitch, LV_STATE_CHECKED);
-    // if (en) {
-    //     label = GuiCreateTextLabel(container, "Pattern Verification");
-    //     imgArrow = GuiCreateImg(container, &imgArrowRight);
-    //     table[0].obj = label;
-    //     table[1].obj = imgArrow;
-    //     table[1].position.x = 396;
-    //     button = GuiCreateButton(container, 456, 84, table, 2, WalletSettingHandler, &walletSetting[2]);
-    //     lv_obj_align(button, LV_ALIGN_DEFAULT, 12, curPositionY);
-    //     curPositionY += 84;
-    //     curPositionY += 12;
-    // }
-
-    lv_obj_t *line = GuiCreateDividerLine(container);
-    lv_obj_align(line, LV_ALIGN_DEFAULT, 0, curPositionY);
-    curPositionY += 12;
-
-    uint8_t fpRegisteredNum = 0;
-    uint8_t textId = 0;
-    for (int i = 0; i < 3; i++) {
-        textId = GetFingerRegisteredStatus(i);
-        if (textId != 0) {
-            printf("text id = %d\n", textId);
-            fpRegisteredNum++;
-            label = GuiCreateTextLabel(container, "");
-            lv_label_set_text_fmt(label, "Finger %d", textId);
-            imgArrow = GuiCreateImg(container, &imgArrowRight);
-
-            table[0].obj = label;
-            table[1].obj = imgArrow;
-            table[1].position.x = 396;
-
-            button = GuiCreateButton(container, 456, 84, table, 2, FingerButtonHandler, &g_curDealFingerIndex[i]);
-            lv_obj_align(button, LV_ALIGN_DEFAULT, 12, curPositionY);
-            curPositionY += 96;
-        }
-    }
-    if (fpRegisteredNum < 3) {
-        label = GuiCreateTextLabel(container, "+ Add Fingerprint");
-        lv_obj_set_style_text_color(label, ORANGE_COLOR, LV_PART_MAIN);
-        table[0].obj = label;
-        button = GuiCreateButton(container, 456, 84, table, 1, WalletSettingHandler, &walletSetting[0]);
-        lv_obj_align(button, LV_ALIGN_DEFAULT, 12, curPositionY);
-    } else {
-        label = GuiCreateNoticeLabel(container, "You can add up to 3 fingerprints");
-        lv_obj_align(label, LV_ALIGN_DEFAULT, 36, curPositionY);
-    }
-}
-
-void GuiWalletFingerDeleteWidget(lv_obj_t *parent)
-{
-    lv_obj_set_style_bg_opa(parent, LV_OPA_0, LV_PART_SCROLLBAR | LV_STATE_SCROLLED);
-    lv_obj_set_style_bg_opa(parent, LV_OPA_0, LV_PART_SCROLLBAR | LV_STATE_DEFAULT);
-
-    lv_obj_t *imgFinger = GuiCreateImg(parent, &imgYellowFinger);
-    lv_obj_align(imgFinger, LV_ALIGN_BOTTOM_MID, 0, -548);
-
-    char buf[32] = {0};
-    sprintf(buf, "Finger %d", GetFingerRegisteredStatus(g_deleteFingerIndex));
-    lv_obj_t *label = GuiCreateLittleTitleLabel(parent, buf);
-    lv_obj_align(label, LV_ALIGN_BOTTOM_MID, 0, -476);
-
-    label = GuiCreateTextLabel(parent, "Remove Fingerprint");
-    lv_obj_set_style_text_color(label, RED_COLOR, LV_PART_MAIN);
-    GuiButton_t table[] = {
-        {
-            .obj = label,
-            .align = LV_ALIGN_BOTTOM_MID,
-            .position = {0, -15},
-        }
-    };
-    lv_obj_t *button = GuiCreateButton(parent, 282, 66, table, NUMBER_OF_ARRAYS(table), FingerDeleteDialogsHandler, &g_deleteFingerIndex);
-    lv_obj_align(button, LV_ALIGN_BOTTOM_MID, 0, -64);
-}
-
-void GuiWalletSetFingerPassCodeWidget(lv_obj_t *parent)
-{
-    static uint32_t walletSetting[2] = {
-        DEVICE_SETTING_FINGER_SETTING,
-        DEVICE_SETTING_RESET_PASSCODE
-    };
-    lv_obj_set_style_bg_opa(parent, LV_OPA_0, LV_PART_SCROLLBAR | LV_STATE_SCROLLED);
-    lv_obj_set_style_bg_opa(parent, LV_OPA_0, LV_PART_SCROLLBAR | LV_STATE_DEFAULT);
-
-    lv_obj_t *label = GuiCreateTextLabel(parent, _("fingerprint_passcode_fingerprint_setting"));
-    lv_obj_t *imgArrow = GuiCreateImg(parent, &imgArrowRight);
-
-    GuiButton_t table[2] = {
-        {
-            .obj = label,
-            .align = LV_ALIGN_DEFAULT,
-            .position = {24, 22},
-        },
-        {
             .obj = imgArrow,
-            .align = LV_ALIGN_DEFAULT,
-            .position = {411, 32},
+            .align = LV_ALIGN_LEFT_MID,
+            .position = {411, 0},
         },
     };
-    lv_obj_t *button = GuiCreateButton(parent, 456, 84, table, NUMBER_OF_ARRAYS(table),
-                                       GuiShowKeyboardHandler, &walletSetting[0]);
+    button = GuiCreateButton(parent, 456, 84, table, NUMBER_OF_ARRAYS(table), OpenSinglePhraseHandler, &walletSetting[0]);
     lv_obj_align(button, LV_ALIGN_DEFAULT, 12, 144 - GUI_MAIN_AREA_OFFSET);
 
-    lv_obj_t *line = GuiCreateDividerLine(parent);
-    lv_obj_align(line, LV_ALIGN_DEFAULT, 0, 240 - GUI_MAIN_AREA_OFFSET);
-
-    label = GuiCreateTextLabel(parent, _("fingerprint_passcode_reset_passcode"));
+    label = GuiCreateTextLabel(parent, _("seed_check_share_phrase"));
     imgArrow = GuiCreateImg(parent, &imgArrowRight);
     table[0].obj = label;
     table[1].obj = imgArrow;
-    button = GuiCreateButton(parent, 456, 84, table, 2, GuiShowKeyboardHandler, &walletSetting[1]);
-    lv_obj_align(button, LV_ALIGN_DEFAULT, 12, 253 - GUI_MAIN_AREA_OFFSET);
+    button = GuiCreateButton(parent, 456, 84, table, NUMBER_OF_ARRAYS(table), OpenSharePhraseHandler, &walletSetting[1]);
+    lv_obj_align(button, LV_ALIGN_DEFAULT, 12, 240 - GUI_MAIN_AREA_OFFSET);
 }
 
-static void FingerButtonHandler(lv_event_t *e)
+void *GuiWalletRecoverySinglePhrase(lv_obj_t *parent, uint8_t wordAmount)
+{
+    lv_keyboard_user_mode_t kbMode = GuiGetMnemonicKbType(wordAmount);
+    g_inputWordsCnt = wordAmount;
+
+    lv_obj_set_style_bg_opa(parent, LV_OPA_0, LV_PART_SCROLLBAR | LV_STATE_SCROLLED);
+    lv_obj_set_style_bg_opa(parent, LV_OPA_0, LV_PART_SCROLLBAR | LV_STATE_DEFAULT);
+
+    lv_obj_t *label = GuiCreateTitleLabel(parent, _("seed_check_single_phrase_title"));
+    lv_obj_align(label, LV_ALIGN_DEFAULT, 36, 12);
+    label = GuiCreateIllustrateLabel(parent, _("seed_check_share_phrase_title"));
+    lv_obj_set_style_text_opa(label, LV_OPA_60, LV_PART_MAIN);
+    lv_obj_align(label, LV_ALIGN_DEFAULT, 36, 72);
+
+    g_recoveryMkb = GuiCreateMnemonicKeyBoard(parent, GuiMnemonicInputHandler, kbMode, NULL);
+    g_recoveryMkb->intputType = MNEMONIC_INPUT_SETTING_VIEW;
+    lv_obj_set_size(g_recoveryMkb->cont, 408, 236);
+    lv_obj_align_to(g_recoveryMkb->cont, label, LV_ALIGN_OUT_BOTTOM_LEFT, 0, 36);
+    lv_btnmatrix_set_selected_btn(g_recoveryMkb->btnm, g_recoveryMkb->currentId);
+
+    lv_obj_t *cont = GuiCreateContainer(lv_obj_get_width(lv_scr_act()), 114);
+    lv_obj_set_align(cont, LV_ALIGN_BOTTOM_MID);
+    lv_obj_set_style_bg_opa(cont, LV_OPA_0, 0);
+    lv_obj_t *btn = GuiCreateBtn(cont, "");
+    lv_obj_t *img = GuiCreateImg(btn, &imgArrowNext);
+    lv_obj_set_align(img, LV_ALIGN_CENTER);
+    lv_obj_align(btn, LV_ALIGN_BOTTOM_RIGHT, -20, -20);
+    g_recoveryMkb->nextButton = btn;
+    lv_obj_add_event_cb(btn, ImportPhraseWordsHandler, LV_EVENT_CLICKED, NULL);
+    // g_recoveryPhraseKb.buttonCont = cont;
+    cont = GuiCreateContainer(lv_obj_get_width(lv_scr_act()), 242);
+    lv_obj_set_align(cont, LV_ALIGN_BOTTOM_MID);
+    lv_obj_set_style_bg_opa(cont, LV_OPA_0, 0);
+    g_recoveryPhraseKb = GuiCreateLetterKeyBoard(cont, NULL, true, g_recoveryMkb);
+    g_recoveryMkb->letterKb = g_recoveryPhraseKb;
+
+    return cont;
+}
+
+// share phrase
+void *GuiWalletRecoverySharePhrase(lv_obj_t *parent, uint8_t wordAmount)
+{
+    g_inputWordsCnt = wordAmount;
+    lv_keyboard_user_mode_t kbMode = GuiGetMnemonicKbType(wordAmount);
+
+    lv_obj_set_style_bg_opa(parent, LV_OPA_0, LV_PART_SCROLLBAR | LV_STATE_SCROLLED);
+    lv_obj_set_style_bg_opa(parent, LV_OPA_0, LV_PART_SCROLLBAR | LV_STATE_DEFAULT);
+
+    g_recoveryMkb = GuiCreateMnemonicKeyBoard(parent, GuiMnemonicInputHandler, kbMode, NULL);
+    g_recoveryMkb->intputType = MNEMONIC_INPUT_SETTING_VIEW;
+    lv_obj_t *label = GuiCreateTitleLabel(parent, _("import_wallet_ssb_title"));
+    lv_obj_align(label, LV_ALIGN_DEFAULT, 36, 12);
+    lv_label_set_recolor(label, true);
+    g_recoveryMkb->titleLabel = label;
+
+    label = GuiCreateIllustrateLabel(parent, _("import_wallet_ssb_desc"));
+    lv_obj_set_style_text_color(label, DARK_GRAY_COLOR, LV_PART_MAIN);
+    lv_obj_align(label, LV_ALIGN_DEFAULT, 36, 72);
+    lv_label_set_recolor(label, true);
+    lv_label_set_text_fmt(label, "Write down the #F5870A %d#-words seed phrase of your first share in the blanks below", g_inputWordsCnt);
+
+    lv_obj_set_size(g_recoveryMkb->cont, 408, 236);
+    lv_obj_align_to(g_recoveryMkb->cont, label, LV_ALIGN_OUT_BOTTOM_LEFT, 0, 36);
+    lv_btnmatrix_set_selected_btn(g_recoveryMkb->btnm, g_recoveryMkb->currentId);
+    lv_label_set_text_fmt(g_recoveryTitle, "%s #F5870A %d#", _("import_wallet_ssb_title"), g_recoveryMkb->currentSlice + 1);
+
+    lv_obj_t *cont = GuiCreateContainer(lv_obj_get_width(lv_scr_act()), 114);
+    lv_obj_set_align(cont, LV_ALIGN_BOTTOM_MID);
+    lv_obj_set_style_bg_opa(cont, LV_OPA_0, 0);
+    lv_obj_t *btn = GuiCreateBtn(cont, "");
+    lv_obj_t *img = GuiCreateImg(btn, &imgArrowNext);
+    lv_obj_set_align(img, LV_ALIGN_CENTER);
+    lv_obj_align(btn, LV_ALIGN_BOTTOM_RIGHT, -20, -20);
+    g_buttonCont = cont;
+    g_recoveryMkb->nextButton = btn;
+    lv_obj_set_style_bg_opa(g_recoveryMkb->nextButton, LV_OPA_30, LV_PART_MAIN);
+    lv_obj_add_event_cb(btn, ImportShareNextSliceHandler, LV_EVENT_CLICKED, NULL);
+
+    cont = GuiCreateContainer(lv_obj_get_width(lv_scr_act()), 242);
+    lv_obj_set_align(cont, LV_ALIGN_BOTTOM_MID);
+    lv_obj_set_style_bg_opa(cont, LV_OPA_0, 0);
+
+    g_recoveryPhraseKb = GuiCreateLetterKeyBoard(cont, NULL, false, g_recoveryMkb);
+    g_recoveryMkb->letterKb = g_recoveryPhraseKb;
+
+    return cont;
+}
+
+static void ImportPhraseWordsHandler(lv_event_t *e)
 {
     lv_event_code_t code = lv_event_get_code(e);
+
     if (code == LV_EVENT_CLICKED) {
-        uint8_t *fingerIndex = lv_event_get_user_data(e);
-        g_deleteFingerIndex = *fingerIndex;
-        uint8_t walletIndex = DEVICE_SETTING_FINGER_DELETE;
-        GuiEmitSignal(SIG_SETUP_VIEW_TILE_NEXT, &walletIndex, sizeof(walletIndex));
+        ImportSinglePhraseWords(g_recoveryMkb, g_recoveryPhraseKb);
     }
 }
 
-static void RecognizeSuccussHandler(lv_timer_t *timer)
-{
-    GUI_DEL_OBJ(g_verifyFingerCont)
-    lv_obj_add_state(g_fpSingerSwitch, LV_STATE_CHECKED);
-    lv_event_send(g_fpSingerSwitch, LV_EVENT_VALUE_CHANGED, NULL);
-    lv_timer_del(g_fpRecognizeTimer);
-    g_fpRecognizeTimer = NULL;
-}
-
-static void RecognizeFailHandler(lv_timer_t *timer)
-{
-    if (g_verifyFingerCont != NULL) {
-        lv_img_set_src(g_imgSignFinger, &imgYellowFinger);
-        lv_obj_add_flag(g_labelSignFailed, LV_OBJ_FLAG_HIDDEN);
-    }
-    lv_timer_del(timer);
-    g_fpRecognizeTimer = NULL;
-}
-
-static void FingerUnlockDeviceHandler(lv_event_t *e)
+static void ImportShareNextSliceHandler(lv_event_t *e)
 {
     lv_event_code_t code = lv_event_get_code(e);
 
     if (code == LV_EVENT_CLICKED) {
-        bool en = lv_obj_has_state(g_fpUnlockSwitch, LV_STATE_CHECKED);
-        if (en) {
-            lv_obj_clear_state(g_fpUnlockSwitch, LV_STATE_CHECKED);
-        } else {
-            lv_obj_add_state(g_fpUnlockSwitch, LV_STATE_CHECKED);
-        }
-        UpdateFingerUnlockFlag(!en);
-        lv_event_send(g_fpUnlockSwitch, LV_EVENT_VALUE_CHANGED, NULL);
-    }
-}
-
-static void FingerDeleteDialogsHandler(lv_event_t *e)
-{
-    lv_event_code_t code = lv_event_get_code(e);
-    if (code == LV_EVENT_CLICKED) {
-        uint8_t *fingerIndex = lv_event_get_user_data(e);
-        char buf[50] = {0};
-        lv_snprintf(buf, sizeof(buf), "Are you sure to remove fingerprint%d?", GetFingerRegisteredStatus(*fingerIndex));
-        lv_obj_t *cont = GuiCreateResultHintbox(lv_scr_act(), 386, &imgWarn, "Remove Fingerprint?",
-                                                buf, "Cancel", DARK_GRAY_COLOR, "Remove", RED_COLOR);
-        lv_obj_t *leftBtn = GuiGetHintBoxLeftBtn(cont);
-        lv_obj_add_event_cb(leftBtn, FingerCancelDeleteHandler, LV_EVENT_CLICKED, cont);
-
-        lv_obj_t *rightBtn = GuiGetHintBoxRightBtn(cont);
-        lv_obj_add_event_cb(rightBtn, FingerDeleteHandler, LV_EVENT_CLICKED, cont);
-    }
-}
-
-static void FingerDeleteHandler(lv_event_t *e)
-{
-    lv_event_code_t code = lv_event_get_code(e);
-    if (code == LV_EVENT_CLICKED) {
-        DeleteFp(g_deleteFingerIndex);
-        lv_obj_del(lv_event_get_user_data(e));
-    }
-}
-
-static void FingerCancelDeleteHandler(lv_event_t *e)
-{
-    lv_event_code_t code = lv_event_get_code(e);
-    if (code == LV_EVENT_CLICKED) {
-        lv_obj_del(lv_event_get_user_data(e));
-        GuiEmitSignal(SIG_SETUP_VIEW_TILE_PREV, NULL, 0);
+        ImportShareNextSlice(g_recoveryMkb, g_recoveryPhraseKb);
     }
 }
