+++ src/ui/gui_widgets/setting/gui_setting_widgets.c
@@ -15,7 +15,6 @@
 #include "bip39.h"
 #include "slip39.h"
 #include "version.h"
-#include "gui_menmonic_test.h"
 #include "presetting.h"
 #include "assert.h"
 #include "gui_qr_hintbox.h"
@@ -30,11 +29,6 @@
 #else
 #include "simulator_model.h"
 #endif
-
-#define DEVICE_SETTING_PASS_MAX_LEN 16
-#define DEVICE_SETTING_RIGHT_LABEL_MAX_LEN DEVICE_SETTING_PASS_MAX_LEN
-#define DEVICE_SETTING_MID_LABEL_MAX_LEN 32
-#define DEVICE_SETTING_LEVEL_MAX 8
 
 typedef void (*setting_update_cb)(void *obj, void *param);
 
@@ -77,23 +71,15 @@ static lv_obj_t *g_resetingCont;               // resetting container
 static lv_obj_t *g_walletSetIcon;              // wallet setting icon
 static lv_obj_t *g_walletSetLabel;             // wallet setting label
 static lv_obj_t *g_mfpLabel;                   // wallet setting label
-static KeyBoard_t *g_recoveryPhraseKb;         // recovery keyboard
-static MnemonicKeyBoard_t *g_recoveryMkb;      // recovery mnemonic keyboard
-static uint8_t g_inputWordsCnt = 0;
 static GuiEnterPasscodeItem_t *g_verifyCode;
 static GuiEnterPasscodeItem_t *g_setPassCode;
 static GuiEnterPasscodeItem_t *g_repeatPassCode;
 static uint8_t g_walletAmount;
-static lv_obj_t *g_recoveryTitle; // recovery title label
-static lv_obj_t *g_buttonCont;    // next buton cont
 
 static lv_obj_t *g_passphraseLearnMoreCont = NULL;
 static KeyboardWidget_t *g_keyboardWidget = NULL;
 
-static void ImportPhraseWordsHandler(lv_event_t *e);
-
 static void OpenPassphraseLearnMoreHandler(lv_event_t *e);
-void RebootHandler(lv_event_t *e);
 void GuiShowKeyboardHandler(lv_event_t *e);
 
 void GuiSettingCloseToTargetTileView(uint8_t targetIndex)
@@ -120,15 +106,6 @@ void GuiSettingAnimSetLabel(const char *text)
     g_waitAnimWidget.label = GuiCreateTextLabel(g_waitAnimWidget.cont, text);
     lv_obj_align(g_waitAnimWidget.label, LV_ALIGN_BOTTOM_MID, 0, -76);
     lv_obj_add_flag(g_waitAnimWidget.cont, LV_OBJ_FLAG_CLICKABLE);
-}
-
-static void ResetClearImportHandler(lv_event_t *e)
-{
-    lv_event_code_t code = lv_event_get_code(e);
-    if (code == LV_EVENT_CLICKED) {
-        ClearMnemonicKeyboard(g_recoveryMkb, &g_recoveryMkb->currentId);
-        GuiClearKeyBoard(g_recoveryPhraseKb);
-    }
 }
 
 static void RecoveryPassphraseHandler(lv_event_t *e)
@@ -922,195 +899,6 @@ void OpenSharePhraseHandler(lv_event_t *e)
     }
 }
 
-// recovery method
-static void GuiWalletRecoveryMethodCheck(lv_obj_t *parent)
-{
-    lv_obj_set_style_bg_opa(parent, LV_OPA_0, LV_PART_SCROLLBAR | LV_STATE_SCROLLED);
-    lv_obj_set_style_bg_opa(parent, LV_OPA_0, LV_PART_SCROLLBAR | LV_STATE_DEFAULT);
-    lv_obj_t *button;
-
-    if (PassphraseExist(GetCurrentAccountIndex()) == true) {
-        lv_obj_t *cont = GuiCreateContainerWithParent(parent, 408, 514);
-        lv_obj_set_style_bg_color(cont, WHITE_COLOR, LV_PART_MAIN);
-        lv_obj_set_style_bg_opa(cont, LV_OPA_10 + LV_OPA_2, LV_PART_MAIN);
-        lv_obj_set_style_radius(cont, 24, LV_PART_MAIN);
-        lv_obj_align(cont, LV_ALIGN_TOP_MID, 0, 156 - GUI_MAIN_AREA_OFFSET);
-
-        lv_obj_t *imgLock = GuiCreateImg(cont, &imgWalletLock);
-        lv_obj_align(imgLock, LV_ALIGN_TOP_MID, 0, 36);
-
-        lv_obj_t *disableTitle = GuiCreateTextLabel(cont, _("Disable Passphrase Wallet"));
-        lv_obj_align(disableTitle, LV_ALIGN_TOP_MID, 0, 140);
-        lv_obj_set_style_text_align(disableTitle, LV_TEXT_ALIGN_CENTER, LV_PART_MAIN);
-
-        lv_obj_t *disableDesc = GuiCreateNoticeLabel(cont, "You are presently using a passphrase-protected wallet. Prior to proceeding with the seed phrase verification process, please restart your device without entering the passphrase.");
-        lv_obj_align(disableDesc, LV_ALIGN_TOP_MID, 0, 188);
-        lv_obj_set_style_text_align(disableDesc, LV_TEXT_ALIGN_CENTER, LV_PART_MAIN);
-        lv_label_set_long_mode(disableDesc, LV_LABEL_LONG_WRAP);
-        lv_obj_set_width(disableDesc, 336);
-
-        lv_obj_t *restartLabel = GuiCreateTextLabel(cont, _("Restart Now"));
-        lv_obj_set_style_text_color(restartLabel, ORANGE_COLOR, LV_PART_MAIN);
-        GuiButton_t restartTable[] = {
-            {
-                .obj = restartLabel,
-                .align = LV_ALIGN_CENTER,
-                .position = {0, 0},
-            },
-        };
-        button = GuiCreateButton(cont, 198, 66, restartTable, NUMBER_OF_ARRAYS(restartTable), RebootHandler, NULL);
-        lv_obj_align(button, LV_ALIGN_BOTTOM_MID, 0, -26);
-        return;
-    }
-    static uint8_t walletSetting[2] = {
-        DEVICE_SETTING_RECOVERY_SINGLE_PHRASE,
-        DEVICE_SETTING_RECOVERY_SHARE_PHRASE
-    };
-
-    lv_obj_t *label = GuiCreateTextLabel(parent, _("seed_check_single_phrase"));
-    lv_obj_t *imgArrow = GuiCreateImg(parent, &imgArrowRight);
-
-    GuiButton_t table[2] = {
-        {
-            .obj = label,
-            .align = LV_ALIGN_LEFT_MID,
-            .position = {24, 0},
-        },
-        {
-            .obj = imgArrow,
-            .align = LV_ALIGN_LEFT_MID,
-            .position = {411, 0},
-        },
-    };
-    button = GuiCreateButton(parent, 456, 84, table, NUMBER_OF_ARRAYS(table), OpenSinglePhraseHandler, &walletSetting[0]);
-    lv_obj_align(button, LV_ALIGN_DEFAULT, 12, 144 - GUI_MAIN_AREA_OFFSET);
-
-    label = GuiCreateTextLabel(parent, _("seed_check_share_phrase"));
-    imgArrow = GuiCreateImg(parent, &imgArrowRight);
-    table[0].obj = label;
-    table[1].obj = imgArrow;
-    button = GuiCreateButton(parent, 456, 84, table, NUMBER_OF_ARRAYS(table), OpenSharePhraseHandler, &walletSetting[1]);
-    lv_obj_align(button, LV_ALIGN_DEFAULT, 12, 240 - GUI_MAIN_AREA_OFFSET);
-}
-
-static void *GuiWalletRecoverySinglePhrase(lv_obj_t *parent, uint8_t wordAmount)
-{
-    lv_keyboard_user_mode_t kbMode = GuiGetMnemonicKbType(wordAmount);
-    g_inputWordsCnt = wordAmount;
-
-    lv_obj_set_style_bg_opa(parent, LV_OPA_0, LV_PART_SCROLLBAR | LV_STATE_SCROLLED);
-    lv_obj_set_style_bg_opa(parent, LV_OPA_0, LV_PART_SCROLLBAR | LV_STATE_DEFAULT);
-
-    lv_obj_t *label = GuiCreateTitleLabel(parent, _("seed_check_single_phrase_title"));
-    lv_obj_align(label, LV_ALIGN_DEFAULT, 36, 12);
-    label = GuiCreateIllustrateLabel(parent, _("seed_check_share_phrase_title"));
-    lv_obj_set_style_text_opa(label, LV_OPA_60, LV_PART_MAIN);
-    lv_obj_align(label, LV_ALIGN_DEFAULT, 36, 72);
-
-    g_recoveryMkb = GuiCreateMnemonicKeyBoard(parent, GuiMnemonicInputHandler, kbMode, NULL);
-    g_recoveryMkb->intputType = MNEMONIC_INPUT_SETTING_VIEW;
-    lv_obj_set_size(g_recoveryMkb->cont, 408, 236);
-    lv_obj_align_to(g_recoveryMkb->cont, label, LV_ALIGN_OUT_BOTTOM_LEFT, 0, 36);
-    lv_btnmatrix_set_selected_btn(g_recoveryMkb->btnm, g_recoveryMkb->currentId);
-
-    lv_obj_t *cont = GuiCreateContainer(lv_obj_get_width(lv_scr_act()), 114);
-    lv_obj_set_align(cont, LV_ALIGN_BOTTOM_MID);
-    lv_obj_set_style_bg_opa(cont, LV_OPA_0, 0);
-    lv_obj_t *btn = GuiCreateBtn(cont, "");
-    lv_obj_t *img = GuiCreateImg(btn, &imgArrowNext);
-    lv_obj_set_align(img, LV_ALIGN_CENTER);
-    lv_obj_align(btn, LV_ALIGN_BOTTOM_RIGHT, -20, -20);
-    g_recoveryMkb->nextButton = btn;
-    lv_obj_add_event_cb(btn, ImportPhraseWordsHandler, LV_EVENT_CLICKED, NULL);
-    // g_recoveryPhraseKb.buttonCont = cont;
-    cont = GuiCreateContainer(lv_obj_get_width(lv_scr_act()), 242);
-    lv_obj_set_align(cont, LV_ALIGN_BOTTOM_MID);
-    lv_obj_set_style_bg_opa(cont, LV_OPA_0, 0);
-    g_recoveryPhraseKb = GuiCreateLetterKeyBoard(cont, NULL, true, g_recoveryMkb);
-    g_recoveryMkb->letterKb = g_recoveryPhraseKb;
-
-    return cont;
-}
-
-void GuiWalletRecoveryDestruct(void *obj, void *param)
-{
-    // todo
-    lv_obj_add_flag(g_recoveryPhraseKb->cont, LV_OBJ_FLAG_HIDDEN);
-    lv_obj_add_flag((lv_obj_t *)obj, LV_OBJ_FLAG_HIDDEN);
-    lv_obj_add_flag(g_buttonCont, LV_OBJ_FLAG_HIDDEN);
-    // lv_obj_del((lv_obj_t *)obj);
-    g_recoveryPhraseKb = NULL;
-    lv_obj_del(g_recoveryMkb->nextButton);
-    g_buttonCont = NULL;
-    // lv_obj_del(obj);
-}
-
-static void ImportPhraseWordsHandler(lv_event_t *e)
-{
-    lv_event_code_t code = lv_event_get_code(e);
-
-    if (code == LV_EVENT_CLICKED) {
-        ImportSinglePhraseWords(g_recoveryMkb, g_recoveryPhraseKb);
-    }
-}
-
-static void ImportShareNextSliceHandler(lv_event_t *e)
-{
-    lv_event_code_t code = lv_event_get_code(e);
-
-    if (code == LV_EVENT_CLICKED) {
-        ImportShareNextSlice(g_recoveryMkb, g_recoveryPhraseKb);
-    }
-}
-
-// share phrase
-static void *GuiWalletRecoverySharePhrase(lv_obj_t *parent, uint8_t wordAmount)
-{
-    g_inputWordsCnt = wordAmount;
-    lv_keyboard_user_mode_t kbMode = GuiGetMnemonicKbType(wordAmount);
-
-    lv_obj_set_style_bg_opa(parent, LV_OPA_0, LV_PART_SCROLLBAR | LV_STATE_SCROLLED);
-    lv_obj_set_style_bg_opa(parent, LV_OPA_0, LV_PART_SCROLLBAR | LV_STATE_DEFAULT);
-
-    g_recoveryMkb = GuiCreateMnemonicKeyBoard(parent, GuiMnemonicInputHandler, kbMode, NULL);
-    g_recoveryMkb->intputType = MNEMONIC_INPUT_SETTING_VIEW;
-    lv_obj_t *label = GuiCreateTitleLabel(parent, _("import_wallet_ssb_title"));
-    lv_obj_align(label, LV_ALIGN_DEFAULT, 36, 12);
-    lv_label_set_recolor(label, true);
-    g_recoveryMkb->titleLabel = label;
-
-    label = GuiCreateIllustrateLabel(parent, _("import_wallet_ssb_desc"));
-    lv_obj_set_style_text_color(label, DARK_GRAY_COLOR, LV_PART_MAIN);
-    lv_obj_align(label, LV_ALIGN_DEFAULT, 36, 72);
-    lv_label_set_recolor(label, true);
-    lv_label_set_text_fmt(label, "Write down the #F5870A %d#-words seed phrase of your first share in the blanks below", g_inputWordsCnt);
-
-    lv_obj_set_size(g_recoveryMkb->cont, 408, 236);
-    lv_obj_align_to(g_recoveryMkb->cont, label, LV_ALIGN_OUT_BOTTOM_LEFT, 0, 36);
-    lv_btnmatrix_set_selected_btn(g_recoveryMkb->btnm, g_recoveryMkb->currentId);
-    lv_label_set_text_fmt(g_recoveryTitle, "%s #F5870A %d#", _("import_wallet_ssb_title"), g_recoveryMkb->currentSlice + 1);
-
-    lv_obj_t *cont = GuiCreateContainer(lv_obj_get_width(lv_scr_act()), 114);
-    lv_obj_set_align(cont, LV_ALIGN_BOTTOM_MID);
-    lv_obj_set_style_bg_opa(cont, LV_OPA_0, 0);
-    lv_obj_t *btn = GuiCreateBtn(cont, "");
-    lv_obj_t *img = GuiCreateImg(btn, &imgArrowNext);
-    lv_obj_set_align(img, LV_ALIGN_CENTER);
-    lv_obj_align(btn, LV_ALIGN_BOTTOM_RIGHT, -20, -20);
-    g_buttonCont = cont;
-    g_recoveryMkb->nextButton = btn;
-    lv_obj_set_style_bg_opa(g_recoveryMkb->nextButton, LV_OPA_30, LV_PART_MAIN);
-    lv_obj_add_event_cb(btn, ImportShareNextSliceHandler, LV_EVENT_CLICKED, NULL);
-
-    cont = GuiCreateContainer(lv_obj_get_width(lv_scr_act()), 242);
-    lv_obj_set_align(cont, LV_ALIGN_BOTTOM_MID);
-    lv_obj_set_style_bg_opa(cont, LV_OPA_0, 0);
-
-    g_recoveryPhraseKb = GuiCreateLetterKeyBoard(cont, NULL, false, g_recoveryMkb);
-    g_recoveryMkb->letterKb = g_recoveryPhraseKb;
-
-    return cont;
-}
 
 static void DelWalletHandler(lv_event_t *e)
 {
@@ -1158,9 +946,7 @@ void GuiWalletRecoveryWriteSe(bool result)
     lv_obj_t *btn;
     lv_obj_t *img;
     if (result) {
-        ClearMnemonicKeyboard(g_recoveryMkb, &g_recoveryMkb->currentId);
-        GuiClearMnemonicKeyBoard(g_recoveryMkb);
-        GuiDelMnemonicKeyBoard(g_recoveryMkb);
+        GuiWalletSeedCheckClearKb();
 
         g_noticeHintBox = GuiCreateHintBox(lv_scr_act(), 480, 356, false);
         img = GuiCreateImg(g_noticeHintBox, &imgSuccess);
@@ -1374,8 +1160,7 @@ void GuiSettingDeInit(void)
     // if (g_recoveryMkb->cont != NULL) {
     //     GUI_DEL_OBJ(g_recoveryMkb->cont)
     // }
-    CLEAR_OBJECT(g_recoveryMkb);
-    GuiMnemonicHintboxClear();
+    GuiWalletSeedCheckClearObject();
     CloseToTargetTileView(g_deviceSetTileView.currentTile, DEVICE_SETTING);
     lv_obj_del(g_deviceSetTileView.tileView);
     lv_obj_del(g_deviceSetTileView.cont);
@@ -1582,7 +1367,7 @@ int8_t GuiDevSettingNextTile(uint8_t tileIndex)
         GuiNvsBarSetRightCb(rightBtn, rightCb, tile);
         break;
     case NVS_BAR_WORD_RESET:
-        rightCb = ResetClearImportHandler;
+        rightCb = ResetSeedCheckImportHandler;
         GuiNvsBarSetRightBtnLabel(NVS_BAR_WORD_RESET, USR_SYMBOL_RESET "Clear");
         GuiNvsBarSetRightCb(NVS_BAR_WORD_RESET, rightCb, NULL);
         break;
@@ -1650,9 +1435,9 @@ int8_t GuiDevSettingPrevTile(uint8_t tileIndex)
         GuiNvsBarSetRightCb(rightBtn, rightCb, g_deviceSettingArray[currentTile].tile);
         break;
     case NVS_BAR_WORD_RESET:
-        rightCb = ResetClearImportHandler;
+        rightCb = ResetSeedCheckImportHandler;
         GuiNvsBarSetRightBtnLabel(NVS_BAR_WORD_RESET, rightLabel);
-        GuiNvsBarSetRightCb(NVS_BAR_WORD_RESET, ResetClearImportHandler, NULL);
+        GuiNvsBarSetRightCb(NVS_BAR_WORD_RESET, ResetSeedCheckImportHandler, NULL);
         break;
     case NVS_RIGHT_BUTTON_BUTT:
         GuiNvsBarSetRightBtnLabel(NVS_BAR_WORD_RESET, "");
